requirements.txt
# MIDI Gateway - GC Lab Chile
# Dependencias Python para emisor y receptor

mido==1.3.0
python-socketio[client]==5.11.0
python-engineio==4.9.0
